@@ -475,37 +521,117 @@
           "exclusiveMinimum": true,
           "scope": "resource",
           "description": "Set the response body size threshold of MB to identify whether a response is a so-called 'large response', only used when 'rest-client.disableHighlightResonseBodyForLargeResponse' and/or 'rest-client.disableAddingHrefLinkForLargeResponse' is set to true"
+        },
+        "rest-client.previewColumn": {
+          "type": "string",
+          "enum": [
+            "current",
+            "beside"
+          ],
+          "default": "beside",
+          "scope": "resource",
+          "description": "Response preview column option. 'current' for previewing in the column of current request file. 'beside' for previewing at the side of the current active column and the side direction depends on 'workbench.editor.openSideBySideDirection' setting, either right or below the current editor column."
+        },
+        "rest-client.previewResponsePanelTakeFocus": {
+          "type": "boolean",
+          "default": true,
+          "scope": "resource",
+          "description": "Preview response panel will take focus after receiving response."
+        },
+        "rest-client.formParamEncodingStrategy": {
+          "type": "string",
+          "enum": [
+            "automatic",
+            "never",
+            "always"
+          ],
+          "default": "automatic",
+          "scope": "resource",
+          "description": "Form param encoding strategy for request body of x-www-form-urlencoded. 'automatic' for detecting encoding or not automatically and do the encoding job if necessary. 'never' for treating provided request body as is, no encoding job will be applied. 'always' for only use for the scenario that 'automatic' option not working properly, e.g., some special characters('+') are not encoded correctly."
+        },
+        "rest-client.addRequestBodyLineIndentationAroundBrackets": {
+          "type": "boolean",
+          "default": true,
+          "scope": "resource",
+          "description": "Add line indentation around brackets('{}', '<>', '[]') in request body when pressing enter."
+        },
+        "rest-client.decodeEscapedUnicodeCharacters": {
+          "type": "boolean",
+          "default": false,
+          "scope": "resource",
+          "description": "Decode escaped unicode characters in response body."
+        },
+        "rest-client.logLevel": {
+          "type": "string",
+          "enum": [
+            "error",
+            "warn",
+            "info",
+            "verbose"
+          ],
+          "default": "error",
+          "scope": "resource",
+          "description": "The verbosity of logging in the REST output panel"
+        },
+        "rest-client.enableSendRequestCodeLens": {
+          "type": "boolean",
+          "default": true,
+          "scope": "resource",
+          "description": "Enable/disable sending request CodeLens in request file"
+        },
+        "rest-client.enableCustomVariableReferencesCodeLens": {
+          "type": "boolean",
+          "default": true,
+          "scope": "resource",
+          "description": "Enable/disable custom variable references CodeLens in request file"
+        }
+      }
+    },
+    "configurationDefaults": {
+      "[http]": {
+        "editor.quickSuggestions": {
+          "other": true,
+          "comments": false,
+          "strings": true
         }
       }
     }
   },
   "scripts": {
-    "vscode:prepublish": "tsc -p ./",
-    "compile": "tsc -watch -p ./",
+    "vscode:prepublish": "webpack --mode production",
+    "webpack": "webpack --mode development",
+    "watch": "webpack --mode development --watch --info-verbosity verbose",
     "postinstall": "node ./node_modules/vscode/bin/install",
     "tslint": "tslint --project tsconfig.json"
   },
   "devDependencies": {
+    "@types/fs-extra": "^5.0.4",
     "@types/mocha": "^2.2.32",
-    "@types/node": "^6.0.40",
+    "@types/node": "^8.10.25",
     "mocha": "^3.2.0",
-    "tslint": "^4.2.0",
-    "typescript": "^2.6.2",
-    "vscode": "^1.0.0"
+    "ts-loader": "^5.3.2",
+    "tslint": "^5.10.0",
+    "typescript": "^3.2.2",
+    "vscode": "^1.0.0",
+    "webpack": "~4.19.1",
+    "webpack-cli": "~3.1.0"
   },
   "dependencies": {
     "adal-node": "^0.1.26",
-    "applicationinsights": "^0.18.0",
-    "autolinker": "^1.4.0",
+    "applicationinsights": "^1.0.5",
     "code-highlight-linenums": "^0.2.1",
     "combined-stream": "^1.0.5",
-    "copy-paste": "^1.3.0",
     "elegant-spinner": "^1.0.1",
     "encodeurl": "^1.0.1",
     "filesize": "^3.3.0",
+    "fs-extra": "^5.0.0",
     "highlight.js": "^9.6.0",
+    "http-proxy-agent": "^2.1.0",
+    "https-proxy-agent": "^2.2.1",
     "httpsnippet": "^1.16.5",
     "iconv-lite": "^0.4.15",
+    "jsonc-parser": "^2.0.2",
+    "jsonpath": "^1.0.0",
     "mime-types": "^2.1.14",
     "moment": "^2.20.1",
     "node-sp-auth": "^2.3.2",
