@@ -1,81 +1,87 @@
 "use strict";
 
-import { window, Uri, workspace } from 'vscode';
-import { ResponseStore } from '../responseStore';
-import { HttpResponse } from '../models/httpResponse';
-import { MimeUtility } from '../mimeUtility';
-import { PersistUtility } from '../persistUtility';
-import { RestClientSettings } from '../models/configurationSettings';
-import { trace } from "../decorator";
-import * as Constants from '../constants';
-import * as fs from 'fs';
-import * as path from 'path';
+import * as fs from 'fs-extra';
 import * as os from 'os';
-
-let cp = require('copy-paste');
-let mime = require('mime-types');
+import * as path from 'path';
+import { Clipboard, env, Uri, window, workspace } from 'vscode';
+import * as Constants from '../common/constants';
+import { HttpResponse } from '../models/httpResponse';
+import { trace } from "../utils/decorator";
+import { MimeUtility } from '../utils/mimeUtility';
+import { PersistUtility } from '../utils/persistUtility';
+import { HttpResponseWebview } from '../views/httpResponseWebview';
 
 export class ResponseController {
-    private _restClientSettings: RestClientSettings;
     public static responseSaveFolderPath: string = path.join(os.homedir(), Constants.ExtensionFolderName, Constants.DefaultResponseDownloadFolderName);
-    public static responseBodySaveFolderPath: string = path.join(os.homedir(), Constants.ExtensionFolderName, Constants.DefaultResponseBodyDownloadFolerName);
+    public static responseBodySaveFolderPath: string = path.join(os.homedir(), Constants.ExtensionFolderName, Constants.DefaultResponseBodyDownloadFolderName);
+    private readonly clipboard: Clipboard;
 
     public constructor() {
-        this._restClientSettings = new RestClientSettings();
+        fs.ensureDir(ResponseController.responseSaveFolderPath);
+        fs.ensureDir(ResponseController.responseBodySaveFolderPath);
+        this.clipboard = env.clipboard;
     }
 
     @trace('Response-Save')
-    public async save(uri: Uri) {
-        if (!uri) {
-            return;
-        }
-        let response = ResponseStore.get(uri.toString());
-        if (response !== undefined) {
-            let fullResponse = this.getFullResponseString(response);
-            let filePath = path.join(ResponseController.responseSaveFolderPath, `Response-${Date.now()}.http`);
+    public async save() {
+        const response = HttpResponseWebview.activePreviewResponse;
+        if (response) {
+            const fullResponse = this.getFullResponseString(response);
+            const defaultFilePath = path.join(ResponseController.responseSaveFolderPath, `Response-${Date.now()}.http`);
             try {
-                await PersistUtility.createFileIfNotExistsAsync(filePath);
-                fs.writeFileSync(filePath, fullResponse);
-                window.showInformationMessage(`Saved to ${filePath}`, { title: 'Open' }, { title: 'Copy Path' }).then(function (btn) {
+                const uri = await window.showSaveDialog({ defaultUri: Uri.file(defaultFilePath) });
+                if (uri) {
+                    let filePath = uri.fsPath;
+                    await PersistUtility.ensureFileAsync(filePath);
+                    await fs.writeFile(filePath, fullResponse);
+                    const btn = await window.showInformationMessage(`Saved to ${filePath}`, { title: 'Open' }, { title: 'Copy Path' });
                     if (btn) {
                         if (btn.title === 'Open') {
                             workspace.openTextDocument(filePath).then(window.showTextDocument);
                         } else if (btn.title === 'Copy Path') {
-                            cp.copy(filePath);
+                            await this.clipboard.writeText(filePath);
                         }
                     }
-                });
-            } catch (error) {
-                window.showErrorMessage(`Failed to save latest response to ${filePath}`);
+                }
+            } catch {
+                window.showErrorMessage('Failed to save latest response to disk.');
             }
         }
     }
 
-    @trace('Response-Save-Body')
-    public async saveBody(uri: Uri) {
-        if (!uri) {
-            return;
+    @trace('Response-Copy-Body')
+    public async copyBody() {
+        const response = HttpResponseWebview.activePreviewResponse;
+        if (response) {
+            await this.clipboard.writeText(response.body);
         }
-        let response = ResponseStore.get(uri.toString());
-        if (response !== undefined) {
-            let contentType = response.getResponseHeaderValue("content-type");
-            let extension = this.getExtension(contentType);
-            let fileName = !extension ? `Response-${Date.now()}` : `Response-${Date.now()}.${extension}`;
-            let filePath = path.join(ResponseController.responseBodySaveFolderPath, fileName);
+    }
+
+    @trace('Response-Save-Body')
+    public async saveBody() {
+        const response = HttpResponseWebview.activePreviewResponse;
+        if (response) {
+            const contentType = response.getHeader("content-type");
+            const extension = MimeUtility.getExtension(contentType, '');
+            const fileName = !extension ? `Response-${Date.now()}` : `Response-${Date.now()}.${extension}`;
+            const defaultFilePath = path.join(ResponseController.responseBodySaveFolderPath, fileName);
             try {
-                await PersistUtility.createFileIfNotExistsAsync(filePath);
-                fs.writeFileSync(filePath, response.bodyStream);
-                window.showInformationMessage(`Saved to ${filePath}`, { title: 'Open' }, { title: 'Copy Path' }).then(function (btn) {
+                const uri = await window.showSaveDialog({ defaultUri: Uri.file(defaultFilePath) });
+                if (uri) {
+                    const filePath = uri.fsPath;
+                    await PersistUtility.ensureFileAsync(filePath);
+                    await fs.writeFile(filePath, response.bodyBuffer);
+                    const btn = await window.showInformationMessage(`Saved to ${filePath}`, { title: 'Open' }, { title: 'Copy Path' });
                     if (btn) {
                         if (btn.title === 'Open') {
                             workspace.openTextDocument(filePath).then(window.showTextDocument);
                         } else if (btn.title === 'Copy Path') {
-                            cp.copy(filePath);
+                            await this.clipboard.writeText(filePath);
                         }
                     }
-                });
-            } catch (error) {
-                window.showErrorMessage(`Failed to save latest response body to ${filePath}`);
+                }
+            } catch {
+                window.showErrorMessage('Failed to save latest response body to disk');
             }
         }
     }
@@ -96,20 +102,5 @@ export class ResponseController {
             body = `${os.EOL}${response.body}`;
         }
         return `${statusLine}${headerString}${body}`;
-    }
-
-    private getExtension(contentType: string): string | boolean {
-        let mimeType = MimeUtility.parse(contentType);
-        let contentTypeWithoutCharsets = `${mimeType.type}${mimeType.suffix}`;
-
-        // Check if user has custom mapping for this content type first
-        if (contentTypeWithoutCharsets in this._restClientSettings.mimeAndFileExtensionMapping) {
-            let ext = this._restClientSettings.mimeAndFileExtensionMapping[contentTypeWithoutCharsets];
-            ext = ext.replace(/^(\.)+/, "");
-            if (ext) {
-                return ext;
-            }
-        }
-        return mime.extension(contentType);
     }
 }
