@@ -1,21 +1,19 @@
 "use strict";
 
-import { window, workspace, OutputChannel } from 'vscode';
-import { PersistUtility } from '../persistUtility';
-import { SerializedHttpRequest } from '../models/httpRequest';
-import { HistoryQuickPickItem } from '../models/historyQuickPickItem';
-import { trace } from "../decorator";
+import * as fs from 'fs-extra';
+import moment from 'moment';
 import { EOL } from 'os';
-import * as fs from 'fs';
-import * as moment from 'moment';
+import { window, workspace } from 'vscode';
+import { Logger } from '../logger';
+import { HistoryQuickPickItem } from '../models/historyQuickPickItem';
+import { SerializedHttpRequest } from '../models/httpRequest';
+import { trace } from "../utils/decorator";
+import { PersistUtility } from '../utils/persistUtility';
 
-let tmp = require('tmp');
+const tmp = require('tmp');
 
 export class HistoryController {
-    private _outputChannel: OutputChannel;
-
-    public constructor() {
-        this._outputChannel = window.createOutputChannel('REST');
+    public constructor(private readonly logger: Logger) {
     }
 
     @trace('History')
@@ -48,7 +46,7 @@ export class HistoryController {
             let document = await workspace.openTextDocument(path);
             window.showTextDocument(document);
         } catch (error) {
-            this.errorHandler(error);
+            this.errorHandler(error, 'Failed to persist the request into history file:');
         }
     }
 
@@ -59,13 +57,13 @@ export class HistoryController {
                 .then(async function (btn) {
                     if (btn) {
                         if (btn.title === 'Yes') {
-                            await PersistUtility.serializeToHistoryFile([]);
+                            await PersistUtility.clearRequests();
                             window.showInformationMessage('Request history has been cleared');
                         }
                     }
                 });
         } catch (error) {
-            this.errorHandler(error);
+            this.errorHandler(error, 'Failed to clear the request history:');
         }
     }
 
@@ -97,13 +95,11 @@ export class HistoryController {
         });
     }
 
-    private errorHandler(error: any) {
-        this._outputChannel.appendLine(error);
-        this._outputChannel.show();
+    private errorHandler(error: any, message: string) {
+        this.logger.error(message, error);
         window.showErrorMessage("There was an error, please view details in output log");
     }
 
     public dispose() {
-        this._outputChannel.dispose();
     }
 }
