@@ -1,6 +1,6 @@
 "use strict";
 
-import { HttpRequest } from '../models/httpRequest';
+import { HttpRequest } from './httpRequest';
 
 export interface IRequestParser {
     parseHttpRequest(requestRawText: string, requestAbsoluteFilePath: string): HttpRequest;
