@@ -1,5 +1,8 @@
 "use strict";
 
+import { ParsedUrlQuery } from 'querystring';
+import { parse as urlParse } from 'url';
+
 export interface HARNameValue {
     name: string;
     value: string;
@@ -25,13 +28,12 @@ export class HARPostData {
     public constructor(public mimeType: string, public text: string) {
         if (mimeType === 'application/x-www-form-urlencoded') {
             if (text) {
-                text = decodeURIComponent(text.replace('+', '%20'));
+                text = decodeURIComponent(text.replace(/\+/g, '%20'));
                 this.params = [];
                 let pairs = text.split('&');
                 pairs.forEach(pair => {
-                    let key: string, value: string;
-                    [key, value] = pair.split('=');
-                    this.params.push(new HARParam(key, value));
+                    let [key, ...values] = pair.split('=');
+                    this.params.push(new HARParam(key, values.join('=')));
                 });
             }
         }
@@ -39,6 +41,23 @@ export class HARPostData {
 }
 
 export class HARHttpRequest {
+    public queryString: HARParam[];
+
     public constructor(public method: string, public url: string, public headers: HARHeader[], public cookies: HARCookie[], public postData: HARPostData) {
+        const queryObj = urlParse(url, true).query;
+        this.queryString = this.flatten(queryObj);
+    }
+
+    private flatten(queryObj: ParsedUrlQuery): HARParam[] {
+        const queryParams: HARParam[] = [];
+        Object.keys(queryObj).forEach(name => {
+            const value = queryObj[name];
+            if (Array.isArray(value)) {
+                queryParams.push(...value.map(v => new HARParam(name, v)));
+            } else {
+                queryParams.push(new HARParam(name, value));
+            }
+        });
+        return queryParams;
     }
 }
