@@ -1,26 +1,22 @@
 "use strict";
 
 import { Stream } from 'stream';
+import { getHeader } from '../utils/misc';
+import { Headers } from './base';
+import { RequestVariableCacheKey } from './requestVariableCacheKey';
 
 export class HttpRequest {
+    public requestVariableCacheKey?: RequestVariableCacheKey;
     public constructor(
         public method: string,
         public url: string,
-        public headers: { [key: string]: string },
+        public headers: Headers,
         public body: string | Stream,
         public rawBody: string) {
     }
 
-    public getRequestHeaderValue(name: string) {
-        if (this.headers) {
-            for (let header in this.headers) {
-                if (header.toLowerCase() === name.toLowerCase()) {
-                    return this.headers[header];
-                }
-            }
-        }
-
-        return null;
+    public getHeader(name: string) {
+        return getHeader(this.headers, name);
     }
 }
 
@@ -28,7 +24,7 @@ export class SerializedHttpRequest {
     public constructor(
         public method: string,
         public url: string,
-        public headers: { [key: string]: string },
+        public headers: Headers,
         public body: string,
         public startTime: number) {
     }
