@@ -1,5 +1,7 @@
 "use strict";
 
+import { getHeader } from '../utils/misc';
+import { Headers } from "./base";
 import { HttpRequest } from "./httpRequest";
 import { HttpResponseTimingPhases } from './httpResponseTimingPhases';
 
@@ -8,26 +10,17 @@ export class HttpResponse {
         public statusCode: number,
         public statusMessage: string,
         public httpVersion: string,
-        public headers: { [key: string]: string },
+        public headers: Headers,
         public body: string,
         public elapsedMillionSeconds: number,
-        public requestUrl: string,
         public bodySizeInBytes: number,
         public headersSizeInBytes: number,
-        public bodyStream: Buffer,
+        public bodyBuffer: Buffer,
         public timingPhases: HttpResponseTimingPhases,
         public request: HttpRequest) {
     }
 
-    public getResponseHeaderValue(name: string) {
-        if (this.headers) {
-            for (let header in this.headers) {
-                if (header.toLowerCase() === name.toLowerCase()) {
-                    return this.headers[header];
-                }
-            }
-        }
-
-        return null;
+    public getHeader(name: string) {
+        return getHeader(this.headers, name);
     }
 }
