@@ -10,14 +10,13 @@ export enum PreviewOption {
 export function fromString(value: string): PreviewOption {
     value = value.toLowerCase();
     switch (value) {
-        case 'full':
-            return PreviewOption.Full;
         case 'headers':
             return PreviewOption.Headers;
         case 'body':
             return PreviewOption.Body;
         case 'exchange':
             return PreviewOption.Exchange;
+        case 'full':
         default:
             return PreviewOption.Full;
     }
