@@ -1,16 +1,19 @@
 "use strict";
 
-import { IRequestParser } from '../models/IRequestParser';
-import { CurlRequestParser } from '../curlRequestParser';
-import { HttpRequestParser } from '../httpRequestParser';
+import { CurlRequestParser } from '../utils/curlRequestParser';
+import { HttpRequestParser } from '../utils/httpRequestParser';
+import { IRequestParser } from './IRequestParser';
 
 export interface IRequestParserFactory {
-    createRequestParser(rawHttpRequest: string);
+    createRequestParser(rawHttpRequest: string): IRequestParser;
 }
 
 export class RequestParserFactory implements IRequestParserFactory {
+
+    private static readonly curlRegex: RegExp = /^\s*curl/i;
+
     public createRequestParser(rawHttpRequest: string): IRequestParser {
-        if (rawHttpRequest.trim().toLowerCase().startsWith('curl'.toLowerCase())) {
+        if (RequestParserFactory.curlRegex.test(rawHttpRequest)) {
             return new CurlRequestParser();
         } else {
             return new HttpRequestParser();
